requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
ortools==9.15.6755
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
supabase==2.10.0
